requirements.txt
requests==0.9.1
python-dotenv==2.20.1